+++ crates/ide-assists/src/handlers/proof_action.rs
@@ -1,18 +1,18 @@
-pub(crate) mod assert_by;
-pub(crate) mod intro_failing_ensures;
-pub(crate) mod intro_failing_requires;
-pub(crate) mod intro_match;
-pub(crate) mod localize_error;
-pub(crate) mod remove_dead_assertion;
+pub(crate) mod insert_assert_by_block;
+pub(crate) mod insert_failing_postcondition;
+pub(crate) mod insert_failing_precondition;
+pub(crate) mod intro_matching_assertions;
+pub(crate) mod decompose_failing_assert;
+pub(crate) mod remove_redundant_assertion;
 pub(crate) mod apply_induction;
-pub(crate) mod wp_move_assertion;
-pub(crate) mod assert_by_reveal;
-pub(crate) mod insert_reveal;
-pub(crate) mod imply_to_if;
+pub(crate) mod weakest_pre_step;
+pub(crate) mod reveal_opaque_in_by_block;
+pub(crate) mod reveal_opaque_above;
+pub(crate) mod convert_imply_to_if;
 pub(crate) mod split_imply_ensures;
 pub(crate) mod intro_forall;
 pub(crate) mod intro_forall_implies;
-pub(crate) mod by_assume_false;
+pub(crate) mod intro_assume_false;
 pub(crate) mod split_smaller_or_equal_to;
 pub(crate) mod seq_index_inbound;
