+++ crates/ide-assists/src/handlers/proof_action/convert_imply_to_if.rs
