+++ crates/ide-assists/src/handlers/proof_action/insert_assert_by_block.rs
