+++ crates/ide-assists/src/handlers/proof_action/insert_failing_precondition.rs
