+++ crates/ide-assists/src/handlers/proof_action/intro_assume_false.rs
