+++ crates/ide-assists/src/handlers/proof_action/intro_matching_assertions.rs
