+++ crates/ide-assists/src/handlers/proof_action/remove_redundant_assertion.rs
