+++ crates/ide-assists/src/handlers/proof_action/reveal_opaque_above.rs
