+++ crates/ide-assists/src/handlers/proof_action/reveal_opaque_in_by_block.rs
