+++ crates/ide-assists/src/lib.rs
@@ -365,21 +365,21 @@ pub(crate) mod handlers {
             // sorted list above?
             //
             // Verus
-            proof_action::assert_by::assert_by,
-            proof_action::intro_failing_ensures::intro_failing_ensures,
-            proof_action::intro_failing_requires::intro_failing_requires,
-            proof_action::intro_match::intro_match,
-            proof_action::wp_move_assertion::wp_move_assertion,
+            proof_action::insert_assert_by_block::assert_by,
+            proof_action::insert_failing_postcondition::intro_failing_ensures,
+            proof_action::insert_failing_precondition::intro_failing_requires,
+            proof_action::intro_matching_assertions::intro_match,
+            proof_action::weakest_pre_step::wp_move_assertion,
             proof_action::apply_induction::apply_induction,
-            proof_action::localize_error::localize_error,
-            proof_action::remove_dead_assertion::remove_dead_assertions,
-            proof_action::assert_by_reveal::assert_by_reveal,
-            proof_action::insert_reveal::insert_reveal,
-            proof_action::imply_to_if::imply_to_if,
+            proof_action::decompose_failing_assert::localize_error,
+            proof_action::remove_redundant_assertion::remove_dead_assertions,
+            proof_action::reveal_opaque_in_by_block::assert_by_reveal,
+            proof_action::reveal_opaque_above::insert_reveal,
+            proof_action::convert_imply_to_if::imply_to_if,
             proof_action::split_imply_ensures::split_imply_ensures,
             proof_action::intro_forall::intro_forall,   
             proof_action::intro_forall_implies::intro_forall_implies,
-            proof_action::by_assume_false::by_assume_false,
+            proof_action::intro_assume_false::by_assume_false,
             proof_action::split_smaller_or_equal_to::split_smaller_or_equal_to,
             proof_action::seq_index_inbound::seq_index_inbound,
         ]
